Rename identity.ts to oktaTokenHelper.ts
